Extracted common methods into filter.py and refactored.
Holds a lot of utility functions, but I only really intend for the
function filter_by_guess to be imported and used. This used to be the
huge chunk of my main method.
